+++ src/Pim/Bundle/EnrichBundle/Form/Type/DatagridConfigurationType.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace Pim\Bundle\EnrichBundle\Form\Type;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolverInterface;
+
+/**
+ * Datagrid configuration form type
+ *
+ * @author    Gildas Quemener <gildas@akeneo.com>
+ * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class DatagridConfigurationType extends AbstractType
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        $builder
+            ->add('columns', 'choice', [
+                'multiple' => true,
+                'expanded' => true,
+                'choices'  => $options['columns'],
+            ])
+            ->add('order', 'hidden');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setDefaultOptions(OptionsResolverInterface $resolver)
+    {
+        $resolver->setDefaults(
+            [
+                'data_class' => 'Pim\Bundle\EnrichBundle\Entity\DatagridConfiguration',
+                'columns'    => [],
+            ]
+        );
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getName()
+    {
+        return 'pim_catalog_datagrid_configuration';
+    }
+}